refactor: implement protocol-based hexagonal architecture
Major architectural improvements:
- Protocol-based cross-module communication (AuthUserProtocol pattern)
- DI as top-level composition root outside hexagonal boundaries
- Value objects for Go principle: accept interfaces, return instances
- Zero cross-module dependencies in domain layers
- Local domain types (AuthUserStatus) to avoid coupling
- Clean API integration with dependency injection
- Eliminated circular dependencies auth.infrastructure ↔ auth.api

Technical changes:
- Move DI from infrastructure to top-level di/ module
- Create auth/domain/protocols.py with protocol definitions
- Create auth/domain/value_objects.py with return types
- Create auth/domain/user_status.py with local enums
- Refactor auth_router.py to use DI container
- Update tach.toml with clean module boundaries
- Streamline documentation to single LLM-optimized guide

Architecture validation:
- All tach boundary checks pass
- Type safety maintained with structural typing
- Auth module serves as reference implementation template
