+++ backend/di/__init__.py
@@ -1,0 +1,7 @@
+"""
+Dependency injection module.
+
+This module serves as the composition root for the application, orchestrating
+all dependencies across different layers of the hexagonal architecture.
+The DI module sits outside the architecture and can depend on all layers.
+"""
