+++ backend/di/user_container.py
@@ -1,0 +1,9 @@
+"""
+Dependency injection containers for user module.
+
+This module provides simple dependency injection functions for creating
+user-related service instances with their dependencies properly configured.
+"""
+
+# Placeholder for future user module DI logic
+# This will be populated when user module DI is needed
